requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
Automat==22.10.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configobj==5.0.8
constantly==23.10.4
datasets==2.21.0
dill==0.3.8
dnspython==2.6.1
dulwich==0.22.1
email_validator==2.2.0
entmax==1.3
evaluate==0.4.1
everett==3.1.0
fastapi==0.112.0
fastapi-cli==0.0.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Jinja2==3.1.4
jsonargparse==3.13.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lightning-utilities==0.11.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.2
portalocker==2.10.1
protobuf==4.25.4
psutil==6.0.0
pyarrow==15.0.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.4.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.18.0
rich==13.7.1
rpds-py==0.20.0
sacrebleu==2.4.2
safetensors==0.4.4
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.13.0
setuptools==72.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.2
tabulate==0.9.0
tokenizers==0.19.1
torch==2.4.0
torchmetrics==0.10.3
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
unbabel-comet==2.2.2
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
wrapt==1.16.0
wurlitzer==3.1.1
xxhash==3.4.1
yarl==1.9.4
zope.interface==6.2
